requirements.txt
langchain
langchain-chroma
langchain-openai
langchain-community
slack-bolt
python-dotenv
chromadb
tiktoken

